+++ src/components/GameTable/GameTable.tsx
@@ -1,15 +1,10 @@
-import { useEffect } from "react";
 import { useAppSelector } from "../../hooks/useAppSelector.hook";
 import TableRow from "./Table/TableRow";
-import { useAppContext } from "../../context/AppContext";
 
 const GameTable = () => {
-  const { handleClickStartGame } = useAppContext();
   const { table } = useAppSelector((state) => state.table);
 
-  useEffect(() => {
-    handleClickStartGame();
-  }, []);
+  
 
   return (
     <section data-game-table className="p-10">