+++ src/components/Options/Options.tsx
@@ -1,34 +1,25 @@
 import { useAppContext } from "../../context/AppContext";
 import { useAppSelector } from "../../hooks/useAppSelector.hook";
-import { useKeydown } from "../../hooks/useKeydown.hook";
 import HowToPlay from "./HowToPlay/HowToPlay";
 import OptionDifficulties from "./OptionDifficulties";
 import OptionsHowTo from "./OptionHowTo";
 import OptionPause from "./OptionPause";
-import { useAppDispatch } from "../../hooks/useAppDispatch.hook";
-import { handlePause } from "../../slices/gameState.slice";
+import OptionLeaderboard from "./OptionLeaderboard";
+import Leaderboard from "./Leaderboard/Leaderboard";
 
 const Options = () => {
-  const { handleClickPauseGame, handleClickHowToPlay } = useAppContext();
-  const { gamePaused, gameLose, howToOpen } = useAppSelector(
+  const { handleClickPauseGame } = useAppContext();
+  const { gamePaused, howToOpen, leaderboardOpen } = useAppSelector(
     (state) => state.gameState,
   );
-  const dispatch = useAppDispatch();
-
-  const handleClickOpenHowToPlay = () => {
-    handleClickHowToPlay();
-    dispatch(handlePause());
-  };
-
-  useKeydown("S", handleClickPauseGame, !gameLose);
-  useKeydown("H", handleClickOpenHowToPlay, !gameLose);
 
   return (
     <section
       data-options
       className="fixed top-4 right-4 flex items-center justify-center gap-2 px-3 py-1.5"
     >
-      <OptionsHowTo isOpen={howToOpen} onClickOpen={handleClickOpenHowToPlay} />
+      <OptionLeaderboard />
+      <OptionsHowTo isOpen={howToOpen} />
       <OptionDifficulties />
       <OptionPause
         isFaded={howToOpen}
@@ -36,6 +27,7 @@ const Options = () => {
         onClickPauseHandle={handleClickPauseGame}
       />
       <HowToPlay isOpen={howToOpen} />
+      <Leaderboard isOpen={leaderboardOpen} />
     </section>
   );
 };