+++ src/main.tsx
@@ -4,7 +4,7 @@ import "./index.css";
 import App from "./App.tsx";
 import AppContextContainer from "./context/AppContext.tsx";
 import { Provider } from "react-redux";
-import { store } from "./store/store.ts";
+import { store } from "./app/store.ts";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>