+++ src/types/leaderboard.ts
@@ -1,0 +1,6 @@
+export type Leaderboard = {
+  date: Date;
+  score: number;
+  difficulty: number;
+  id: string;
+}